Do not abort polling when instances do not start instantly
The current flow expected the instances to start basically in the poll delay,
if that didn't happen the wait was aborted. If no new instances are found an
additional check will look for the actual deployment that was started, if this
deployment is still running the polling continues until the timeout.

If instances are not staged instantly because of a tense resource situation
the deployment does not fail anymore.